requirements.txt
gspread=6.2.1
google-auth=2.41.1
python-dotenv=1.2.1
fastapi=0.121.0
uvicorn=0.38.0
pytest=9.0.0